Create README.md
he City of New York, is the most populous city in the United States. It is diverse and is the financial capital of USA. It is multicultural. It provides lot of business opportunities and business friendly environment. It has attracted many different players into the market. It is a global hub of business and commerce. The city is a major center for banking and finance, retailing, world trade, transportation, tourism, real estate, new media, traditional media, advertising, legal services, accountancy, insurance, theater, fashion, and the arts in the United States. This also means that the market is highly competitive. As it is highly developed city so cost of doing business is also one of the highest. Thus, any new business venture or expansion needs to be analyzed carefully. The insights derived from analysis will give good understanding of the business environment which help in strategically targeting the market. This will help in reduction of risk. And the Return on Investment will be reasonable. The basis of this study is to find the best location for investors to open a new pub in NYC city. Sales in the drinking place sector in the United States have steadily grown since the mid 1990's reaching 23.15 billion U.S. dollars in 2015. The sector includes bars, pubs, lounges, taverns and nightclubs as well as other drinking places that primarily sell alcoholic beverages for immediate consumption. The National Restaurant Association estimated that food and drink sales in U.S. bars and taverns would reach 19.9 billion U.S. dollars in 2016. This is a relatively small figure in comparison to the wider restaurant industry, for which food and drink sales were expected to rise to approximately 783 billion U.S. dollars during the same year. New York is a major metropolitan area with more than 8.4 million (Quick Facts, 2018) people living within city limits.


With these criteria given by the investing group, based on previous success in other markets, the objective is to locate and recommend to the investors, the target audience, which neighborhood(s) of NYC will be the best choice to start growth plan. The information gained will assist in choosing the right location by providing data about the population of each neighborhood, in addition to other established venues present in these areas.This report explores which neighborhoods and boroughs of New York City have the most as well as the pubs. Additionally, I will attempt to answer the questions “Where should I open an pubs?” and “Where I can find best pub?”